Creating the save method in the ComputerRepository
