+++ Program.cs
@@ -1,6 +1,8 @@
 ﻿using Microsoft.Data.Sqlite;
 using LabManager.Database;
 using LabManager.Repositories;
+using LabManager.Models;
+
 
 var databaseConfig = new DatabaseConfig();
 var databaseSetup = new DatabaseSetup(databaseConfig);
@@ -26,17 +28,8 @@ if(modelName == "Computer")
         string ram = args[3];
         string processor = args[4];
 
-        var connection = new SqliteConnection("Data Source=database.db");
-        connection.Open();
-        var command = connection.CreateCommand();
-
-        command.CommandText = "INSERT INTO Computers VALUES($id, $ram, $processor)";
-        command.Parameters.AddWithValue("$id", id);
-        command.Parameters.AddWithValue("$ram", ram);
-        command.Parameters.AddWithValue("$processor", processor);
-
-        command.ExecuteNonQuery();
-        connection.Close();
+        var computer = new Computer(id, ram, processor);
+        computerRepository.Save(computer);
     }
 }
 if(modelName == "Lab")
@@ -79,7 +72,7 @@ if(modelName == "Lab")
         command.Parameters.AddWithValue("$block", block);
 
         command.ExecuteNonQuery();
-        
+
         connection.Close();
     }
 }
